pyproject.toml
[tool.poetry]
name = "pieface"
version = "0.1.0"
description = ""
authors = ["scarter13 <scarter13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-mimeparse = "^1.6.0"
django-tastypie = "^0.14.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
